backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
supabase
python-dotenv
websockets

# Testing
pytest
pytest-asyncio
httpx
